Possibility to declare function visibility
It often happens, the assembly functions are internal
to the implementation and should not be accessible
outside of DSO.

This patch adds a possibility to declare a function
as hidden. That way it won't get added to the dynamic
symbol table and hence it won't get exported outside
of shared library.